# docs/openapi.yaml
openapi: 3.0.3
info:
  title: FinCommerce Auth & Device API
  description: API specifications for registration, login, multi-factor verification, and session management in FinCommerce.
  version: 1.0.0
servers:
  - url: http://localhost:8000/api/v1
paths:
  /auth/register:
    post:
      summary: Register a new user account
      description: Validates phone/email uniqueness and audits password against public leak databases.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - full_name
                - email
                - phone
                - password
              properties:
                full_name:
                  type: string
                  example: Somchai Dev
                email:
                  type: string
                  format: email
                  example: somchai@gmail.com
                phone:
                  type: string
                  example: "0812345678"
                password:
                  type: string
                  format: password
                  example: P@ssw0rdStrength!
      responses:
        '201':
          description: User registered successfully.
        '400':
          description: Invalid input parameters or leaked password flagged.
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
                    example: "Password chosen was found in previous public breaches. Please choose a secure password."

  /auth/login:
    post:
      summary: Authenticate user
      description: Checks user credentials and evaluates adaptive risks.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - password
              properties:
                username:
                  type: string
                  description: Email or mobile number
                  example: somchai@gmail.com
                password:
                  type: string
                  format: password
                  example: P@ssw0rdStrength!
                remember_me:
                  type: boolean
                  default: false
      responses:
        '200':
          description: Login successful. Returns JWT token or triggers MFA challenge.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [SUCCESS, MFA_REQUIRED]
                  token:
                    type: string
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                  mfa_token:
                    type: string
                    example: mfa_challenge_uuid_here
        '401':
          description: Invalid credentials.
        '403':
          description: Account locked due to rate limiting.

  /auth/mfa/verify:
    post:
      summary: Verify MFA OTP
      description: Validates the 6-digit OTP code sent via SMS.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - mfa_token
                - otp_code
              properties:
                mfa_token:
                  type: string
                  example: mfa_challenge_uuid_here
                otp_code:
                  type: string
                  pattern: "^[0-9]{6}$"
                  example: "123456"
      responses:
        '200':
          description: OTP verified. Returns final session access token.
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        '400':
          description: Invalid or expired OTP.

  /devices:
    get:
      summary: Get all active sessions for the user
      description: Lists details of all connected browsers/devices.
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of active sessions.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    session_id:
                      type: string
                      format: uuid
                    device_name:
                      type: string
                      example: "Safari on macOS (14.2)"
                    ip_address:
                      type: string
                      example: "182.52.120.44"
                    location:
                      type: string
                      example: "Bangkok, Thailand"
                    last_active:
                      type: string
                      format: date-time

  /devices/revoke/{session_id}:
    post:
      summary: Revoke session access
      description: Terminates the session ID, logging out the corresponding device.
      security:
        - BearerAuth: []
      parameters:
        - name: session_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Session revoked successfully.
        '401':
          description: Unauthorized.
        '404':
          description: Session ID not found.

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT